页面布局优化； 接口数据调整； 代码逻辑优化；

+++ app/src/main/java/com/example/juanshichang/bean/CartBean.kt
@@ -44,7 +44,7 @@ class CartBean : Serializable{
         var code: String = "",
         var sort_order: String = "",
         var title: String = "",
-        var value: Int = 0
+        var value:  Double=0.0
     )
 
     data class TotalX(
